Upstream Cobalt changes after rebasing v8
Cobalt rebased to V8 8.8 a few months ago. Here are some new changes we
introduced during the rebase. Some changes could be helpful to upstream
V8.
One thing to point out is that memcpy is implemented in a universal way
so we moved it out of wrappers.h. On the contrary, strdup got in.

Change-Id: I173710af51fcfe1673def37983de9862fb863e5a