[d8/infra] Add flag to help diagnose Mac arm64 test flakiness
Add a --reject-rosetta flag to d8 that bails out early if it detects
it's running on MacOS with Rosetta process translation. This is to
test the hunch that, for unknown reasons, d8 is erroneously running
under Rosetta on the Mac arm64 debug bot.

The flag is disabled by default, but this CL also modifies the build
config to enable it on the troubled bot.

Bug: chromium:1210489
Change-Id: I63c9b1a25caa3ae6f4dc9a780bd6bb01c3e655f3