[sparkplug] Fix enable batch compilation per default
Remove the neg-neg implication sparkplug --> baseline-batch-compilation,
because it is wrong in the current setting.
Since sparkplug is off per default, the implication will turn off batch
compilation.
When sparkplug is turned on explicitly, there is no implication to turn
on batch compilation again.
Since batch compilation is gated behind --sparkplug anyways we can
safely remove it.

Bug: v8:11790
Change-Id: I8f5ffb542625bc8061ceef02bae688edecea8438