[heap] Use snapshot for Code objects in concurrent marker
Allow updating code objects from the main thread. The concurrent marker
now needs to implement our synchronization protocol also for code
objects. This means that outgoing references are stored into a snapshot
and only if marking the object black succeeds, only then those
references in the snapshot are guaranteed to be valid and marked.

Bug: v8:11872
Change-Id: I2fff10694824a1da5a0ac223026fce6bf9841a2e