[compiler] Skip MapRef descriptor array's serialize methods
We don't need to serialize the descriptor array when concurrent
inlining so we can skip the methods related to serializing.

While doing so we bumped into two broken tests in serializer-tester.
Since we are planning on removing the serializer soon, it makes more
sense to remove them than trying to fix them.

Bug: v8:7790
Change-Id: Ib1feea16f09edebbd9927afc4178ed10f85fd42f