[compiler] Fix serialization for Function#bind
It was not in sync with the optimization, which relies on
inspecting up the length and name fields even for bound
functions.

To make a now meaningful serializer test actually pass, I have
to to make some changes to the test setup.

I'm moving also moving the function name and length index
constants from JSFunction to JSFunctionOrBoundFunction for
clarity.

Change-Id: I36dd3c80996ccb53810c7ea9bfceb5c84ffd60ab