[sparkplug] Trigger compilation only when we don't have baseline data
If we trigger baseline compilation unconditionally on bytecode budget
interrupt after we have compiled the function with sparkplug already, we
will arm back-edges for OSR unconditionally. Since the OSR arming
mechanism is the same for Sparkplug and TurboFan, this means that we
will immediately arm back-edges for OSR to TurboFanned code.

Bug: v8:11656
Change-Id: Ic8e5d1899a3a203321e063ada474d39f2627ea88