[fastcall] Set no JS execution flag from TurboFan
This CL implements setting the JAVASCRIPT_EXECUTION_ASSERT
on the isolate from generated code, so we don't need to create
an expensive class in the embedder callback.

Bug: chromium:1218898
Change-Id: Ia05b49281ab4c1cc3ac34caf2dfadb79feb86e84


Migrate PerIsolateAssertScope storage to boolean[]

This CL modifies the underlying storage of PerIsolateAssertScope from
a bitfield to a boolean array. This slightly increases the space taken
by the isolate, but allows for easier access to the individual fields,
which is a prerequisite for implementing assertion scopes in TurboFan.

Bug: chromium:1218898
Change-Id: Ia5e43352ebba28be6f013376b75f13ec8d5dc972