[heap] Tie process-wide CodeRange lifetime to any remaining Heaps
Currently the process-wide CodeRange, once created, lives until process
shutdown. This CL changes it to be alive as long as there is a Heap,
when the last Heap is gone it gets destroyed and will be recreated the
next time a Heap is created. This behavior is shared with
SingleCopyReadOnlyArtifacts.

Bug: v8:11929
Change-Id: I8a545926c3a4122991f9682bd3fd90e72697ea5a