Merged: [compiler] Representation mismatch detection missed a case
In GetOwnFastDataPropertyFromHeap, we read a property value
then compare it with the expected representation. We already
had code to bail out of that particular optimization if there
was a mismatch, however it missed the case of expected
representation being a HeapObject, and when a Smi value was found.

The fix is to use the excellent pre-existing method
Object::FitsRepresentation() to make this check. Thusly, all
cases are handled.

Bug: chromium:1225607, v8:7790
(cherry picked from commit 758816f4388704d82442769163b371c1b64aac86)
No-Try: true
No-Presubmit: true
No-Tree-Checks: true

Change-Id: I7564762db19dbbdb7156b8cca6b574e68f4cc490