[Intl] throw instead of DCHECK while long locale
This is a temp fix to thorw instead of DCHECK in debug build.
The correct fix depends on the landing of
https://github.com/unicode-org/icu/pull/1762

Once that land I will cherrypick into chrome to fix the function correctly.
But the current (before this CL) behavior is not harmful in release build.
It basically does not do the max nor min just return itself.

Bug: chrome:1224869
Change-Id: Iebce2ab0a5ce047e83e8fce05db8290212e64509