[inspector] Provide function name for breakpoints.
When setting (or resolving) a breakpoint, have the inspector also send
the name of the function in which the breakpoint was set (in addition
to the exact location within the script). The idea is to show this in
the Breakpoints view in the DevTools front-end, since for developers
knowing the name of the functio is equally important (if not more
important) than knowing the exact file and line.

Bug: chromium:700516
Change-Id: I968a4dc6949f03703308acd1c10bd41ac8d50895