[compiler] Remove MapData::SerializeElementsKindGeneralizations
.. and the corresponding map cache.

This cache was only used for CreateArray reductions, thus we mostly
expect to see unmodified JSArray initial maps. These are already
accessible from the native context and don't need to be cached
separately.

We may also see initial maps for custom JSArray subclasses. New map
allocation may be necessary in this case if the requested elements
kind differs from that of the current given initial map. Since we
can't (easily) allocate on the background thread, we skip the
optimization starting with this CL.

Bug: v8:7790
Change-Id: Ib8d81094e1572f49eda18e9ec485d317cec62473