[heap] Guard unprotected memory chunk with a mutex
Use a mutex guard when the unprotection is triggered from a compaction
space in which case it is actually parallel.

Main-thread only unprotection does not require acquiring the mutex.

The list itself is only used from the main thread and thus the actual
process does not require a mutex.

The issue was introduced in https://crrev.com/c/2966382

Bug: v8:11982
Change-Id: I593c0659eb5a96c8206d0b4014f07ab13827be85