[sparkplug] On deopts, defer decision to execute baseline / bytecode
This is in preparation for baseline code flushing. After a deopt we
choose to execute baseline / bytecode based on whether SharedFunctionInfo
has any baseline code. With baseline code flushing, it is possible that
baseline code is flushed after this point and before we start executing
the unoptimized code (for ex: materializing objects). To handle such
situations this CL updates the BaselineEnterAt* builtins to check for
baseline code and restart either at baseline / bytecode.

Bug:v8:11947
Change-Id: I2084e38196c882f802d1186ff8c9ab881a35b16b