[web snapshot] Support classes
This CL adds support for classes with methods.

A new ValueSerializer is added and classes are serialized separetely
from functions, although the common parts are handled in the same way
and abstracted away. The function prototype is serialized as an
object and any missing information is set up again during
deserialization. FunctionFlagsToFunctionKinds() is updated to allow for
more function kinds. Lastly, context serialization is updated to support
serializing BlockContexts and creating ScopeInfos of type CLASS_SCOPE.

Bug: v8:11525, v8:11706
Change-Id: I16ca7cbc17b1811721081cda05124ce36073f9be