[bigint] Inline fast-path String-to-BigInt conversion
This significantly speeds up parsing of small BigInts.
The full mechanism relies on std::vector, which uses C++ heap
allocations that are relatively expensive compared to the overall
effort required for parsing small BigInts. This patch adds a fast
path that stack-allocates required data structures for inputs up
to a certain threshold.
Additionally, this patch changes the conversion of characters to
their numeric values from computations to a lookup table, which is
a bit faster for radixes <= 10 (where the first attempted range was
always successful) and a lot faster for radixes > 10.
Additionally, this patch forces the {ConsumeChar} method, which was
always intended to get inlined, to actually get inlined.

Bug: v8:11515
Change-Id: Ifd8ec4799ac34447ba6d4350b7788b559307784c