Explicitly include windows.h in simulator-arm64.cc
Now that V8 doesn't globally include windows.h we need to explicitly
include it in source files that need it. This fixes a Windows ARM64
build break triggered by crrev.com/c/3042215.

I did a test build of ARM64 V8 with and without this patch in order to
reproduce the errors and confirm that there are no others.

Bug: chromium:796644, chromium:1237067
Change-Id: I43045fcaab9e6296629775a9705aae6b2066adef